requirements.txt
# Prophet EPL Dependencies
numpy>=1.21.0
scipy>=1.7.0
requests>=2.25.0
python-dotenv>=0.19.0

# Optional (for testing)
pytest>=7.0.0
